+++ packages/core/tests/e2e/next-server/CHANGELOG.md
@@ -4,5 +4,59 @@
 
 ### Patch Changes
 
+- Updated dependencies [008725ec]
+- Updated dependencies [520fb2d5]
+  - ai@3.2.1
+
+## null
+
+### Patch Changes
+
+- Updated dependencies [b965dd2d]
+- Updated dependencies [85ef6d18]
+  - ai@3.2.0
+
+## null
+
+### Patch Changes
+
+- Updated dependencies [85712895]
+  - ai@3.1.37
+
+## null
+
+### Patch Changes
+
+- Updated dependencies [4728c37f]
+- Updated dependencies [8c49166e]
+  - ai@3.1.36
+
+## null
+
+### Patch Changes
+
+- Updated dependencies [06123501]
+  - ai@3.1.35
+
+## null
+
+### Patch Changes
+
+- Updated dependencies [d25566ac]
+- Updated dependencies [87a5d27e]
+  - ai@3.1.34
+
+## null
+
+### Patch Changes
+
+- Updated dependencies [6fb14b5d]
+- Updated dependencies [05536768]
+  - ai@3.1.33
+
+## null
+
+### Patch Changes
+
 - Updated dependencies [3cabf078]
   - ai@3.1.32