+++ packages/mistral/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @ai-sdk/mistral
 
+## 0.0.18
+
+### Patch Changes
+
+- Updated dependencies [85712895]
+- Updated dependencies [85712895]
+  - @ai-sdk/provider-utils@0.0.15
+
+## 0.0.17
+
+### Patch Changes
+
+- 4728c37f: feat (core): add text embedding model support to provider registry
+- 7910ae84: feat (providers): support custom fetch implementations
+- Updated dependencies [7910ae84]
+  - @ai-sdk/provider-utils@0.0.14
+
 ## 0.0.16
 
 ### Patch Changes