+++ packages/react/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @ai-sdk/react
 
+## 0.0.4
+
+### Patch Changes
+
+- 008725ec: feat (@ai-sdk/react): add experimental_useObject to @ai-sdk/react
+- Updated dependencies [008725ec]
+  - @ai-sdk/ui-utils@0.0.4
+
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies [85712895]
+- Updated dependencies [85712895]
+  - @ai-sdk/provider-utils@0.0.15
+  - @ai-sdk/ui-utils@0.0.3
+
+## 0.0.2
+
+### Patch Changes
+
+- Updated dependencies [7910ae84]
+  - @ai-sdk/provider-utils@0.0.14
+  - @ai-sdk/ui-utils@0.0.2
+
 ## 0.0.1
 
 ### Patch Changes