+++ packages/solid/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @ai-sdk/solid
 
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies [008725ec]
+  - @ai-sdk/ui-utils@0.0.4
+
+## 0.0.3
+
+### Patch Changes
+
+- @ai-sdk/ui-utils@0.0.3
+
+## 0.0.2
+
+### Patch Changes
+
+- @ai-sdk/ui-utils@0.0.2
+
 ## 0.0.1
 
 ### Patch Changes